과정 등록/삭제/수정 API 구현

+++ src/main/java/com/mathflat/parkdingco/service/CreateCourseService.java
@@ -8,5 +8,5 @@ public interface CreateCourseService {
 
     void createBy(Subject newSubject);
 
-    void create(Long studentId, Long subjectId, Long score);
+    void create(Long studentId, Long subjectId, Integer score);
 }
